Rotharium Platform
Explainder video for the Rotharium Platform. Rotharium Platform is the gate to the blockchain and enables you to integrate your apps into the blockchain ecosystem. Rotharium Platform helps to create decentralized and transparent business processes. Rotharium Platform interfaces allow connection of standard cloud or on-premise database solutions to the blockchain.